refactor(test): MeshWeaver.PathResolution.Test drops MeshWeaver.AI — guard included
Two thread cases move to MeshWeaver.AI.Test as ThreadAddressResolutionTest. Their fixture is
NOT incidental: they build AI.Thread content under a _Thread satellite path and the shape
being resolved IS that. Substituting a non-AI node type would have changed the scenario, so
they travel with the engine. The resolution mechanism itself is core and its coverage stays
in place — 136/136 there without the reference.

🚨 The ProjectReference was never a code dependency. NoStaticCollectionsTest enumerates
MeshWeaver.*.dll from AppContext.BaseDirectory at RUNTIME, so it only ever covers the
assemblies in the running test's bin — and the reference existed purely to put AI.dll where
the guard would find it. Dropping it naively leaves the guard GREEN having stopped looking
at AI, which is coverage lost with no signal.

So the guard follows its subject: one linked <Compile> into MeshWeaver.AI.Test, no
duplicated logic, and it travels with that project when the engine leaves.

Linking it proved itself immediately, twice. It ran and FAILED — first because the four
MeshWeaver.AI.* allow entries had been removed on the wrong assumption that AI would no
longer be scanned anywhere (restored: they are legitimate CONST allowances needed wherever
AI.dll is scanned), then because AI.Test's larger dependency closure brought six assemblies
into scope for the FIRST time. A no-op link would have sat silently green.

Those six are allow-listed, each verified rather than assumed: every one is a
declaration-with-initializer whose only later uses are reads (.Contains, lookups) — the
immutable read-only lookups NoStaticState.md permits. They were surfaced by WIDENED
COVERAGE, not by new code, and the entries say so.

Note this widens the guard repo-wide, so this PR classifies six statics unrelated to AI. The
narrower option was to record a coverage gap the way #2293 did for the portal hosts' module
declarations; coverage won because leaving AI unguarded is the failure this whole exercise
exists to avoid, and the six were trivially verifiable.

Verified: both projects build -c Release -warnaserror at 0 Error(s); PathResolution.Test
136/136; AI.Test 4/4 including the linked guard.

No What's New entry: test-only refactor.

Refs #2276
