+++ test/MeshWeaver.PathResolution.Test/AddressResolutionTest.cs
@@ -2,7 +2,6 @@
 using System.Threading.Tasks;
 using System.Reactive.Threading.Tasks;
 using System.Reactive.Linq;
-using MeshWeaver.AI;
 using MeshWeaver.Hosting.Monolith.TestBase;
 using MeshWeaver.Mesh;
 using MeshWeaver.Mesh.Services;
@@ -22,8 +21,9 @@ public class AddressResolutionTest(ITestOutputHelper output) : MonolithMeshTestB
     private const string AppPath = "app";
 
     protected override MeshBuilder ConfigureMesh(MeshBuilder builder)
+        // The thread cases that needed AddAI() moved to MeshWeaver.AI.Test.ThreadAddressResolutionTest
+        // (#2276); what remains resolves plain node paths and needs no AI types.
         => base.ConfigureMesh(builder)
-            .AddAI()
             .AddSampleUsers();
 
     private async Task EnsureNodesCreated()
@@ -226,90 +226,7 @@ public class AddressResolutionTest(ITestOutputHelper output) : MonolithMeshTestB
 
     #region ThreadMessage Path Resolution
 
-    /// <summary>
-    /// Creates a Thread with a ThreadMessage child and verifies path resolution
-    /// resolves the full ThreadMessage path (not the Thread path with remainder).
-    /// </summary>
-    [Fact(Timeout = 30000)]
-    public async Task ResolvePath_ThreadMessageNode_ResolvesToFullMessagePath()
-    {
-        // Create Thread node
-        var threadPath = "User/Roland/_Thread/test-resolution-1234";
-        var threadNode = new MeshNode("test-resolution-1234", "User/Roland/_Thread")
-        {
-            Name = "Test Resolution",
-            NodeType = ThreadNodeType.NodeType,
-            Content = new AI.Thread
-            {
-                Messages = ["msg1"]
-            }
-        };
-        await NodeFactory.CreateNode(threadNode).Should().Emit();
 
-        // Create ThreadMessage child node
-        var msgNode = new MeshNode("msg1", threadPath)
-        {
-            NodeType = ThreadMessageNodeType.NodeType,
-            Order = 1,
-            Content = new ThreadMessage
-            {
-                Role = "user",
-                Text = "Hello",
-                Timestamp = System.DateTime.UtcNow,
-                Type = ThreadMessageType.ExecutedInput
-            }
-        };
-        await NodeFactory.CreateNode(msgNode).Should().Emit();
-
-        // Resolve the ThreadMessage path — should return the full message path, no remainder
-        var resolution = await PathResolver.ResolvePath($"{threadPath}/msg1").Should().Emit();
-
-        resolution.Should().NotBeNull("ThreadMessage node exists at {0}/msg1", threadPath);
-        resolution!.Prefix.Should().Be($"{threadPath}/msg1",
-            "should resolve to the full ThreadMessage path, not the Thread path with remainder");
-        resolution.Remainder.Should().BeNull("exact match should have no remainder");
-    }
-
-    /// <summary>
-    /// Verifies that the Thread node itself still resolves correctly
-    /// when ThreadMessage children exist.
-    /// </summary>
-    [Fact(Timeout = 30000)]
-    public async Task ResolvePath_ThreadNode_ResolvesCorrectlyWithChildren()
-    {
-        var threadPath = "User/Roland/_Thread/test-parent-5678";
-        var threadNode = new MeshNode("test-parent-5678", "User/Roland/_Thread")
-        {
-            Name = "Test Parent",
-            NodeType = ThreadNodeType.NodeType,
-            Content = new AI.Thread
-            {
-                Messages = ["m1"]
-            }
-        };
-        await NodeFactory.CreateNode(threadNode).Should().Emit();
-
-        var msgNode = new MeshNode("m1", threadPath)
-        {
-            NodeType = ThreadMessageNodeType.NodeType,
-            Order = 1,
-            Content = new ThreadMessage
-            {
-                Role = "assistant",
-                Text = "Response",
-                Timestamp = System.DateTime.UtcNow,
-                Type = ThreadMessageType.AgentResponse
-            }
-        };
-        await NodeFactory.CreateNode(msgNode).Should().Emit();
-
-        // Resolve the Thread path — should match the Thread node exactly
-        var resolution = await PathResolver.ResolvePath(threadPath).Should().Emit();
-
-        resolution.Should().NotBeNull();
-        resolution!.Prefix.Should().Be(threadPath);
-        resolution.Remainder.Should().BeNull();
-    }
 
     #endregion
 }
